% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{klog_create_text}
\alias{klog_create_text}
\title{What is there here?}
\usage{
klog_create_text(dt)
}
\arguments{
\item{dt}{data frame containing the keylogs.}
}
\value{
a tibble object containing the running text.
}
\description{
Returns a dataset with one row for each individual key that was typed.
Information about mouse clicks are not included. Key states refer to the
state when the key was pressed. Times are all recentered to the start of
the data. Useful for studing individual keys and bursts of activity.
}
\examples{
example_path <- system.file("keylogs-1621971722509.csv", package = "keylog")
dt <- klog_readfile(example_path)
klog_create_text(dt)
}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
