% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{klog_events}
\alias{klog_events}
\title{Augement keylog data to the level of individual keys.}
\usage{
klog_events(dt, rename = TRUE)
}
\arguments{
\item{dt}{data frame containing the keylogs.}
}
\value{
a tibble object containing the key-based data.
}
\description{
Returns a dataset with one row for each individual key that was typed.
Information about mouse clicks are not included. Key states refer to the
state when the key was pressed. Times are all recentered to the start of
the data. Useful for studing individual keys and bursts of activity.
}
\examples{
example_path <- system.file("keylogs-1622057428556.csv", package = "keylog")
dt <- klog_readfile(example_path)
klog_events(dt)
}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
