% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{klog_readfile}
\alias{klog_readfile}
\title{Read a CSV file containing keylog data.}
\usage{
klog_readfile(input)
}
\arguments{
\item{input}{path to the csv file containing the data.}
}
\value{
a tibble object containing the keylog data.
}
\description{
The file is a standard CSV file and can generally be read with existing
methods. This function simply makes sure that the correct column types
are selected when loading the file and fills in the spaces and new line
characaters correctly.
}
\examples{
example_path <- system.file("keylogs-1622057428556.csv", package = "keylog")
klog_readfile(example_path)

}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
